+++ ErsatzTV.Infrastructure/Data/Configurations/Collection/JellyfinCollectionConfiguration.cs
@@ -6,5 +6,18 @@ namespace ErsatzTV.Infrastructure.Data.Configurations;
 
 public class JellyfinCollectionConfiguration : IEntityTypeConfiguration<JellyfinCollection>
 {
-    public void Configure(EntityTypeBuilder<JellyfinCollection> builder) => builder.ToTable("JellyfinCollection");
+    public void Configure(EntityTypeBuilder<JellyfinCollection> builder)
+    {
+        builder.ToTable("JellyfinCollection");
+
+        builder.Property(c => c.Etag)
+            .HasMaxLength(36)
+            .IsUnicode(false);
+
+        builder.Property(c => c.ItemId)
+            .HasMaxLength(36)
+            .IsUnicode(false);
+
+        builder.HasIndex(c => c.ItemId);
+    }
 }