+++ ErsatzTV.Infrastructure/Data/Configurations/MediaItem/JellyfinSeasonConfiguration.cs
@@ -6,5 +6,18 @@ namespace ErsatzTV.Infrastructure.Data.Configurations;
 
 public class JellyfinSeasonConfiguration : IEntityTypeConfiguration<JellyfinSeason>
 {
-    public void Configure(EntityTypeBuilder<JellyfinSeason> builder) => builder.ToTable("JellyfinSeason");
+    public void Configure(EntityTypeBuilder<JellyfinSeason> builder)
+    {
+        builder.ToTable("JellyfinSeason");
+
+        builder.Property(s => s.Etag)
+            .HasMaxLength(36)
+            .IsUnicode(false);
+
+        builder.Property(s => s.ItemId)
+            .HasMaxLength(36)
+            .IsUnicode(false);
+
+        builder.HasIndex(s => s.ItemId);
+    }
 }